+++ src/Mesh/CartesianFace.h
@@ -80,7 +80,7 @@ public:
 			else
 				assert(false);
 			//             Number  x1    y1    x2    y2 IsDomainBoundary IsRemovedOnCoarserGrid
-			fprintf(file, "%lu %.17g %.17g %.17g %.17g %d %d\n", this->Number, _shape.Origin->X, _shape.Origin->Y, p2.X, p2.Y, this->IsDomainBoundary, this->IsRemovedOnCoarserGrid);
+			fprintf(file, "%d %.17g %.17g %.17g %.17g %d %d\n", static_cast<int>(this->Number), _shape.Origin->X, _shape.Origin->Y, p2.X, p2.Y, this->IsDomainBoundary, this->IsRemovedOnCoarserGrid);
 		}
 		else
 			assert(false);
